RTS-8153 Check user's main account when synchronising areas
If Sebra gives us a stale (non main) user account for a person as
staff on an area, the wrong one would be shown in employee lists.

This commit checks each username and substitutes stale ones for the
main user name where necessary. Note that this will slow down the
synchronisation of areas quite a bit.

Also includes an improvement that stops unnecessary staff list saves.