RTS-8153 Check stale users during Sebra synchronisation
Adds user checks for staff and content managers during area
synchronisation with Sebra. Users that does not exist in Sebra,
are listed with their non-main user account, lack main user in
Sebra, or are missing or blocked in w3 - will be listed on the
area node in the field_uib_staff_status.

The commit also adds adding/removing of user aliases to imple-
mentations of hook_user_insert(), and hook_user_update().