RTS-8197 - H2 tabs scroll up on page load
Removing two lines with scroll to top and changing the time the
page should continously scroll to the top while the page is
loading.

This should make the page seem less intrusive regarding the time
the page is unable to scroll.

Pages with these H2 tabs should now only scroll to the top once
and not lock the page to being "unscrollable".