RTS-7693 Added new module UiB Content List to features
The new content type workflow is made a little more generic as it is
implemented as a content list with a choice of display style. Only one
display style is currently available (workflow).

The content type is basically a listing of content and displayed one
item at a time with a common title, lead and a progress bar. The
current step (node) is displayed as a full node view.

The content type is added to features.