RTS-8272 Display disc in bullet lists on profile pages
The style-list was set to its initial value in the "current project
block" and in the container for tabs (work, publications, etc). But
set to none again for publications (both from cristin and selected
by users) since they contain lists created by the system that should
maintain their previous style. In short the places where the user is
in control of content now display discs in bullet lists.