RTS-8387 Add new taxonomy and module uib_geography for countries
Adds a new module uib_geography that handles a new hierarchical
taxonomy vocab uib_geography. This contains the names of countries
and continents.

The data are picked up from FS via their web service for countries,
and it is checked for iso conformity with the Drupal functions.

Note that discrepancies are logged, but not fixed. Thus we get the
same continents and country names as in FS. Norwegian and English
country names are put in a translatable field.

The commit also contains a drush command for synchronising the
country taxonomy with FS.