RTS-8294 New article without area creates error page
The reason was that, before saving, the system checks if users of
level 2 and 3 are area managers. This check was done even if the
area was empty creating an error during a SELECT query to the DB.

The system now checks if there is an area assigned before trying to
determine if the user has a right to add articles to this area.
Even though the validation process could be stopped at that moment,
it was decided to let the process continue so that all missing
required fields are treated in the same manner.