RTS-8356 Make sure address overrides are registered
Today only the fields used by the sebra sync will be altered if a
user has registered an override for an address field. And the
override is only registered if the address 1 field (thoroughfare)
is one of the altered fields. By checking for empty fields in the
alternate address array and swapping the entire address array if the
number of altered fields is larger than 1 (country will always have a
country code), we can be sure all overrides are caught.