RTS-6923 Adding unpublished content in related content
Added the possibility to add unpublished content to articles.

In article view, unpublished article teasers will show up with the
node-unpublished default background color.

When adding a new related article, it will be displayed if the
related content is published or not.

Still an issue: If an article is saved, there will not be any
indication whether or not the related content is unpublished
in the edit view.