RTS-6378 External content should be visible in slideshow
Added filter to let users add external content to the slideshow
window if there is a media file.

When trying to show external content in the slideshow view we need to
make the view aware of the difference in data structure between an
article and external content. There is a difference both in what
image type that is chosen and what text field is used. By adding the
external content field as substitute values for the article fields on
no results behaviour we get the desired result.