RTS-7224 Create more space for text in slideshow
With long titles and leads the text tend to expand beyond the
transparent overlay on the main image. By refactoring the slideshow
view and tweaking the output based on article beeing an event or not
we are able to use the full width of the container to display title
and lead for articles of type news or infopage. Another minor tweak is
to reduce to margin between title and lead to 10px.