RTS-8518 Rewrite output of views to point edit/delete to correct lang
The urls for the generated edit and delete buttons in views in the
webdesk gets the language prefix from the current language rather
than the language of the nodes attached. By rewriting the output of
the view we are able to alter the language prefix when the urls prefix
is different to the nodes language.