RTS-8793 Static links on exchange agreement page
The exchange agreements use the uib_study content type, and it
already includes an editable links collection. This commit adds
a number of static links in a hard coded block that will be
displayed in the same place as the link section - but only for
uib_study nodes of type "exchange".

If a user adds links to the links section, in addition, those
will be displayed above the static links.