RTS-8803 View page for exchange agreements
Contains a view with simple exposed filters, a list of agreements
grouped by countries, and a bit of styling to make it look like
the other study views. Also, an update task for a Norwegian alias.

In an uib_study implementation of hook_views_pre_render(), we
setup a results text, and fix the number of displayed exchange
agreements for each country.