RTS-8677 Using a filter to display the jobbnorge feed in infopage
We are going to use an infopage to display the feed of vacant
positions at uib. To achieve this we use a text filter that expands
to a rendered version of the feed from jobbnorge. New pre and post
rendering functions are added to achieve the wanted markup.

Basic styling are added in the sass module vacancies.