RTS-8956 Remove HTML markup from translation strings
The recommended way to have a text with links inside is to add the link tag
<a href> within the text in the t() function. However the translation
interface does not allow to save html tags.

The current solution changes the text so that the link is at the end, leaving
the text within the t function and then concatinating the link using l().