RTS-8935 Remove label for related persons in uib_content_list steps
When rendering a node as a full page we do a lot of modifications to
the ouput in template.php. When a full node is rendered as a step in
a content list the output do not get the modifications we do in
template.php. We have do insert the modifications in the
content_list module.