RTS-8935 Made content list use the newslist view
We want the same information displayed for our users when adding a step
in the content list as they get when adding related content to an
article. To supply this service we make a new display in the newslist
view that is a clone of the view used by articles but with the content
type content list removed.

It should only be possible to add articles of type info pages to the
steps of content list. The view display is restricted to showing this
kind of content.