RTS-8805 Added exchange agreements to exchange tab in study programme
If a study programme has exchange agreements that is facilitated as a
part of the programme the agreements should show in the exchange tab.
The agreements should show with the name of the place of study and a
link to the study report tab in the agreement page.

For now the agreement title is shown since we do not save the name of
the institution in w3 (for now anyway).