RTS-8968 Simplify fact box content for network exchange agreements
The following info is not displayed for NORDPLUS or else when there
are more than two parties to the agreement: Country, town, web site,
application code, number of student exchange places, and the fact
box foot-note which was connected to the application code.

In addition, It fixes a typo that stopped the application code from
being displayed.