RTS-8943 Article metadata should show in steps in content lists
The article metadata (created and updated) are processed in
template.php and are not displayed when a node is rendered as a step
in a content list.

By moving our code that processes the metadata from template.php to
a function in the article module we may reuse the same code to
process and insert the metadata when needed.