RTS-8981 Use FS tag studienivaintervallnavn for study level
The study level for exchange agreements was initially picked up
from the study level of a list of connected study programmes.
However, although not incorrect, SA preferred to supply the text
string themselves.
Note that this text string is (seemingly) supplied in a Norwegian
language, which makes it impractical in terms of translation,
because w3 uses English as default language for translations.