RTS-8803 Add link to exchange agreement exposed filter description
It was possible to add a link to the description, not by using the
regular views exposed filter description - but the one provided by
the better exposed filters module. It allowed embedded html in the
text.

The commit also includes two small changes to the css. One for the
link in the text, and the media query width. The filters are only
displyed in one row on desktop size screens (from 960px). This is
because at that point, other media queries (such as in the infamous
responsive-sidebars) decreases the width of the page. Keeping the
filters in a single row at that point would mess up the filter
display.