RTS-9098 Form fields needs to be hidden by restricting access
Form fields in study nodes of type exchange were hidden from users
of various levels by using hide. This caused the values in the
fields to be erased when the nodes were saved. Some of these values
were used to filter nodes in the exchange-agreement view. Hiding
the fields in the form by access restriction makes the problem go
away.