RTS-8951 Added redirect
Added redirect to stop users from endring up in hybrid pages
where the language is something inbetween norwegian and
english. Users entering hybrid pages will be redirected to
the language of the original article.

As a side effect. If the article doesn't have a translation,
users will be unable to change language