RTS-9172 Remove locality information from network exch. agreements
When an exchange agreement is a network, notably a Nordplus agree-
ment, the information about the institution and city should not
show.

The removal had to be implemented in two places since the data is
added to the block "study_content", the corresponding tabs links
later.

This commit also fixes a problem with translating a info group
header to css id: Previously only a few non-ascii characters were
handled, resulting in utf-8 characters in the id names when there
were non-Norwegian characters on the info group header.