RTS-9308 Not only bachelor programmes should show exch. agreements
Showing a list of exchange agreements on the corresponding selected
study programmes did not work for programmes other than bachelor
programmes.

The reason was that the only FS infotype group code implemented
was "ba_utveksling" - only relevant to bachelor programmes.

The solution checks for group codes ending in "_utveksling". This
is common to all the relevant group codes according to Sveinjarle
Nymark.