RTS-9321 Fixed bug in slideshow view
After the release of R19 an article with relevant links put in the
slideshow  will link to the url given for the first item in the
relevant links section. This is due to the fact that the same field
is used in relevant links in articles and url in external content.

To avoid this we use another level of rewriting so that only items
without the field article type uses the rewriting supposed to
address external content.