RTS-9116 Upgrade to eu_cookie_compliance 1.14
The module has changed its machine name from eu-cookie-compliance
to eu_cookie_compliance this implementation updates drupal.make to
reflect this change.

In addition 2 more array elements are added to the variables
eu_cookie_compliance_en and eu_cookie_compliance_nb that represent
the texts that will be shown to the user in the Yes/No buttons.