RTS-9316 Drush script to add and update uib_organisation taxonomy
Creates/updates the uib_organisation vocabulary with data from
the uib_ou and area nodes. Basically, there is a taxonomy term
for each ou node for each faculty and department/institute
(including unit) and one at the top for the university. It is
hierarchical, based on place codes (ou id).

The names are picked up from the area nodes where possible, else
from the ou.