RTS-9316 Synchronise studies with new uib_organisation vocab
The taxonomy vocabulary is setup to be synchronised right after
the Sebra places. This is because the vocabulary is strongly
dependent upon the uib_ou.

The study synchronisation is modified to also add/update study
connection to the uib_organisation vocabulary in the
field_uib_study_organisation.

A first synchronisation is setup to run in bin/post-upgrade.