RTS-9459 Related content on Employee Pages to show Faculty OU
The code now looks for the "parent" or faculty OU, the one whose code
finishes in 0000. So only images of faculty OU are shown.

During testing of this issue another problem was found. Different
types of content use different media fields, this commit addresses
that problem also.

uib_content_list however replaces the media in template.php which
happens after the hook used in this solution.