RTS-9564 Modify geography vocabulary term names
In order to use the geography vocab as autocomplete exposed filter
in view form for both English and Norwegian, we would need to have
both languages in the term name - since that is what the autocom-
plese seems hard-wired to use.

The commit includes a script that modifies existing term names,
and a corresponding change to the synchronisation drush script.