RTS-9122 Area aliases in both English and Norwegian
The existing code for maintaining area aliases (i.e. the set of
aliases corresponding to various area centered views etc.), was
modified to handle different aliases for English and Norwegian.

The code picks up the area node language and sets aliases
accordingly. The default language is used in case no language is
set.

A post-upgrade script is added for a one-time change of aliases.
The script does not handle redirects from the old aliases since
this is done automatically with the current config.