RTS-9305 Upgrade to addressfield-7.x-1.0
The new version of addressfield makes some of the fields of an address
required by default if there is a country selected. This commits changes
the default configuration to make all fields optional (for uib_user). No
changes where made to uib_area (ou) but the manner used by the module to
store the configuration has chenges. This upgrade requires drush updatedb.