RTS-9564 Filter the country filter on exchange agreements view page
One side-effect when switching to the "has taxonomy term (with depth)"
filter in the exchange_agreements view is related to a number of
network agreements that are connected to several countries. Filtering
for one country would pull up all countries such agreements are
connected to. This would be unexpected for the user, and this commit
adds a manual filter that removes those extra hits.