RTS-3311 Images on newspage nedd to be clickable
This commit adds the module linked_field which provides the functionality
of adding a link to a field in the manage display settings of the field.
This allows adding a link to the fields in the correponding views and
displays.

This commit modiefies the teaser display of the uib_article content and the
frontpage_profiled_articles view to satisfy the requisits of the issue.