RTS-9283 Read student reports link on exchange agreements view page
The commit adds a direct link to student reports from the
exchange agreements view page. It is exactly like what was
removed in issue rts/9168, with the addition of only showing the
link if there are published reports for that exchange agreement.

The presence if reports are checked by using a new api function
at the student reports site.

Note that just like the display of student reports in the first
place, the functionality added here is hard coded as considering
all reports to be Norwegian.