RTS-9835 Move "studieløp" tab in studies next to/after "oppbygging"
This commit creates an auxiliary array pre_tab_items that
can be rearranged after all the elments are all created but before
the tab headers are added to the render array. It also moves the
code that generates the construction table to its own function so
it is easier to see how the tabs are created.