BS-307305 Apply ldap patch 1775658
This is patch currently only required for the tk.app.uib.no application, but
since it reuses the drupal repo that w3 creates we include the patch here.
This isn't an issue in w3 because we have not enabled the ldap_profile module.

This was a problem for tk because it imports users from Sebra instead of
letting the ldap module create them on first login.  This worked at first
but after we changed to use the uib unames as users.name and enabled the
ldap_profile to sync the full_name field we ran into this error when
somebody with no earlier ldap history tried to log in.

https://www.drupal.org/node/1775658