RTS-7927 Filter special courses views on faculties (and semester)
Filtering these views utilised the uib_organisation taxonomy, limited
to only faculties.

In order to filter on semester we use two new course labels; "spring"
and "autumn". FS-pres will set these up from FS-data and the labels
will be synchronised automatically. The two label taxonomy terms were
added manually in the admin interface in prod in order to be able to
use the real taxonomy term ids needed for the limited "has taxonomy
term" filter.

The translation of the better exposed filters is handled by a
pre-existing function uib_study__translate_exposed_filter(). This was
modified here, but may have grown out of its usefulness.

I used a variable for translation in the t() function. The Drupal api
function description says this should be avoided. In this case, the
variable would not really contain user data, and it was cleaned by
passing the string through the check_plain() function. Without this,
it would have been difficult to translate the labels.