RTS-9797 Filter modifications for Exchange agreements view
The Faculty/departments filter is remodeled into a with depth
filter for faculties only. Departments are not shown (but inclu-
ded in the results). The filter is no longer collapsible. Also,
the "University of Bergen" was included, and code added for this
particular selection not to be "with depth" - else all agreements
would have shown when checking UiB.

This commit also hides all agreements if no filter is in use.