RTS-7225 New content type: views page
In order to be able to have an editable text field on top of a view,
this content type is added.

It has fields for area connection, title, info-text, and the
machine name of a views block that is rendered below the node
contents.

The node title field, and the info-text are translatable.