RTS-9797 Keep listing agreements initially
Removing all agreements from being listed initially casued a problem
in issue #7225 when converting the view to displaying a block instead
of a page: The block became empty and nothing (not filters either)
was shown. This might be expected since there is no reason to filter
when there are no results to filter.
Thus, in this commit we refrain from emptying the list initially.