Field instances feature update related to linked fields upgrade
This is most probably related to the upgrade of the linked fields
module (commit e579e67982 in the Drupal repo). These changes have
been appearing when performing feature update, but since they are
unrelated to any rts issue, the changes have not been commited
before. Here they are all in their glory.