RTS-7225 course and study programme views modified
Moved the "all courses" and all "study programmes" displays of the
course view into a new view "study_blocks". They used to be pages
and resided in the already bloated Study view (in the uib_area
feature).

Set these displays up as blocks to be displayed by uib_views_pages.
Modified styling, and had to add a template file in order to move
the "Found xxx courses" text below the filters.

Translation of the level filter did not work automatically.