RTS-10408 Manually translate study category filter
For the views of all courses and all study programmes, the transla-
tion of the exposed filter of study category stopped working when
the view displays were converted to blocks instead of pages in its
own view.

The translation is now performed manually similarly to how it is
done for other exposed filters: text replacement in the html code
as returned by the "better exposed filters" module.