RTS-8802 Select semester based on query parameter
Added URL query to have the semester chooser working without
JavaScript. Added the PHP code that will make this work.

Downside is that the page will now loaded completly every time a
semester is changed.  Upside is you can link directly to a semester

(This commit squashes together a modified version of the
earlier reverted commits for this issue.)