RTS-10503 The url-filter removed too much
Issue #9026 changed the url-filter for related links.
In cases described in that issue (news article), the
language indicator appeared twice and code was added
to remove those in order to get a functional url.
This does not seem to work now, since one language
indicator would be expected.

In the links described in this issue (phD press release
article) we do not get a double language indicator. Too much
of the url was removed, and the code would not take into
account that file urls should not have language.