RTS-10674 Introduce the calendar3 module
This is yet another attempt at rewriting the calendar views for the
site.  The new approach uses object oriented db_select() to create the query
and JS to toggle views based on selected event types.

This baseline also explores a new presentation where we try to
use font-awsome icons to represent the event types.