RTS-10633 Place construction table
The rule now is that the construction table is inserted right
after the last occurence of an 'presentasjon' or 'oppbygging'
group.  These are usually in the front.

This modifies the approach taken in RTS-9835 for inserting the
construction table in the right location.  Now we just maintain
an index where the table is to be inserted.