RTS-10906 Make sure staff list from Sebra is always an array
Getting the staff list for an area from Sebra ("visninger"),
results in different structures depending on the number of
users in the list: empty, a string when there is one user, or an
array when there are more than one users. This seems to be caused
by the simplexml_load_string() PHP function, and our code expected
it to be an array. The solution is to make sure we have an array.