RTS-10831 Add support for ResearchGate on user profile
Adds the possibility to add researchgate in field_uib_user_social
_media on the user profile. An svg icon was created, as Research
Gate could not supply one.

The previous svg logos are displayed with a small margin, and this
does not. I have been unable to find out why - but it means that
this logo is slightly larger than the others. It fills the 32px
given to the uib-social-media field-item in the css (while the
previous icons do not).