RTS-9280 Robustify Sebra synchronisation
When the Sebra web service fails to respond, and this happens
regularly, the area and place synchronisations fail. The user
synchronisation may fail a single user at a time.

Tests indicate that even a single retry may help. This commit
adds a retry (and logging) functionality when accessing the
Sebra web service. It is configured to retry three times before
giving up, retrying after 2, 4 and 6 seconds.