RTS-10912 Set connected content to mandatory in content list
It was possible to add a step title, but no connected content for a
step in a content list. This would result in an error.

By setting the connected content field in a step to be mandatory,
it is no longer possible to try to save a step with title but no
content connection. Thus, the error is avoided.