RTS-11257 Prevent the new theme crashing the site
Some tests done in template.php in the uib_zen theme should also be
done in the uib_study module to prevent other themes from crashing
the site. With this commit applied we are also able to set Bartik
or Garland as default theme and not bring the site down.