RTS-10980 Expansion filter for press release list
Adding (*pressrelease*) to article text will expand a list of the
last 10 press releases published (uib_article of type press_release).

The list of press releases will be displayed in a rendered block,
sorted (descending) with the newest at the top. Separate lists
(<ul>) for each date with the date as <h3> header.