RTS-11319 Added html.tpl.php and page.tpl.php file
To override Drupal's default doctype of XHTML 1.0 we add a html.tpl.php
that uses our preferred doctype of html 5.

By adding a page.tpl.php file we get the basic structure of our site the
way we want it and get rid of a lot of the default divs drupal provide.