RTS-11330 Added settings and logo file for using a custom logo
By default drupal searches for a file named logo.png in the theme
folder. We want to use a svg file so we need to add the logo file and
settings in the .info file to let drupal know what file to look for.

We should also add the settings to the uib_zen.info file and remove the
settings from features.