RTS-11313 Calendar title modification
The issue was to make the calendar title "full width", i.e. except
for a rather wide left margin where the year and month takes up
some space. The suggested layout is only relevant to large
screens.

In order to place the calendar title in the grid and style it,
it was moved into its own div with a "calendar-header" class.