RTS-11278 Structural re-make of the calendar3 page
The date picker width was set to 17em by the date module (as well
as the jquery css). This made it hard to style the column. By
using 10 instead of 5 grid columns, several things could be
achieved: (1) A reasonable width (3 grid columns) of the right
column which made it possible to set the enclosed elements to
100%. (2) Column 1 could be empty which would stop the year and
month to slide off the screen: they would stay in column 1.

The break point for screen size switch was increased from 720 to
740. Else, the date picker table and its header would change in
an unlinked way between 720 and 740 px screen size.