RTS-11329 The study_related block should only be inserted in uib_study
On our staging and production servers the php version is 5.3. In this
php version drupal crashes when it tries to access non-existent fields.
Area and article does not contain field_uib_study_part_of and the
attempt to access this field causes the crash.

This is the patch Gisle supplied in this issue.