RTS-11413 Added code to create and update aliases for inner market nodes
The nodes of type inner market does not have neither language nor area
set. This makes it hard to get a useful path alias from the module
pathauto.

The solution proposed in this commit is to use hook_node_insert and
hook_node_update to create aliases for norwegian and english on creating
and updating node of this type.