RTS-11930 Tweaking the scheduler settings
We want the users only to be able to pick every full hour as a value to
avoid confusion and by changing the date popup minute increment to 60
this is achieved. Altering the time format to display only hours and
minutes will probably also avoid confusion. Finally it is now possible
to ommit the time. A default time at 00:00:00 is then entered. This is
possible for a user to alter later.