RTS-11414 Added function for registration of article with filter
To avoid unecessary administration of nodes where uib_field_text
contains a filter a function for automatic registration of nodes is
created. The function tries to match the patterns set up for our text
filters and if found registers the node id in an array if the node id
does not exist in the array.

As a first step registration of nodes using the pressrelease filter is
added, but it should be easy to expand the function to register other
filters if needed.