RTS-12069 Adding typekit support to the new theme
We want to use the default fonts of our graphic profile in the new
theme. To achieve this goal we need to add support for typekit as
myriad and garamond are both adobe fonts and are only available for use
on the web through a typekit account.

The necessary javascripts are added by drupal_add_js in
hook_preprocess_html.