RTS-12422 Making an article of type infopage ready for new theme
For an article of type info page the author should not be displayed in
the article info. We only want to display the publish date and the date
of the last update to the article.

Related content, relevant links, related persons and file attachments
is moved from content to content_bottom. This is probably a change
needed for all article types.

Switch from using unset() to hide() to stop fields from rendering.