RTS-12344 Output exhibition section below the regular events
It's easier to output this as part of the same block as the regular
events.  The previous version based on views made it two separate
blocks.  I don't think we need the flexibility to move the blocks
to separate regions.

We don't replicate the parentisis around the date ranges of the
exhibition entries as requested in RTS-12176.