RTS-10978 Log in as admin while updating areas
The node translation sync will only update nodes that the current user
has access to view.  This caused the synced node translations for area
nodes not to spread to unpublished areas, causing havoc.  We can address
this by running the area sync as the 'admin' user.

By default drush runs as the anonymous user.  It's possible to provide
the --user=1 option to drush for commands that are bootstraped to the
LOGIN phase, but I found no good way to make it default that way, so
we'll just keep bootstraping to FULL and then do the login manually.