Replace nb_NO collation on OS X
OS X provides uses "no_NO" as locale instead of "nb_NO" (and "nn_NO") as
you find on most other systems.  We used to hack around this by creating
a symlink in /usr/share/locale, but with El Capitan's SIP this is no
longer possible.  Postgres will inherit these as its exposed collations.

Another approach I tried was to set up an collate alias using
'CREATE COLLATION "nb_NO" FROM "no_NO"'.  This kind of works, but then
you need to get rid of the "pg_catalog." prefix from the dumps, so it's
actually less effort to just replace the collation id in the dumps.

It would be better to have the condition depend on the OS where the
database is hosted, but for now just assume it's hosted locally.