RTS-14310 Markup and styling needed for user profile layout
To get the wanted layout for the top of the user profile we need to be
able to use flexbox. This means we need a div around all the image and
all other fields above the contact information. We also needs a div
around all the other fields except the user image. This will allow use
to create a flexbox container with two flex items.

Also added surronding div to the user-contact-info

This commit also contains the styling done in this issue in SG