RTS-14626 Fixed bug with selected publications
The way selected publications was added in template.php was depending on
a user to also have content in field_uib_user_field. As these fields
should be rendered as tabs it was inserted at the wrong position in the
markup.

This fix aims to remove this dependency and insert selected publications into
field_uib_user_field before the tabs are rendered.