RTS-14977 Max-heigt for media thumbnails in editor mode
On inserting and editing images we want the images used to have a max
height set to 200px. After the upgrade of the media module the images
was show in full size.

To achive this we had added some extra css in the uib_admin module and
made sure the media browser used our custom css.