RTS-15052 Added uib_ckeditor_config module
We need a module to enable us to modify the settings in ckeditor beyond
what the wysiwyg module offers in the gui. To achieve this we need to
tell ckeditor to load a custom config file through this module. Custom
settings can be added to this javascript file and will be picked up by
ckeditor.