RTS-15171 Updated autocomplete view for profiled articles
In the uib_zen the profiled articles was rendered through a view in a
slideshow. All nodes that filled certain criteria where allowed to add
to the field, but items without main media was not displayed.

In uib_w3 the profiled articles is rendered as a field an the added
articles are all displayed. We need to alter the autocomplete view
(newslist) to prevent nodes without media attached to turn up. As we do
this we also wants to accept small media.

This commit alters the view to only return nodes with main media or
small media. Other filter criteria stays the same.