RTS-14473 Views pages needs titles
We generate a lot of view pages for areas. These pages needs to get a
proper title. The problem is that we remove the title in the page
template and insert it in the correct place in template.php.By hooking into
theme_hook_suggestion and current area we are able to generate the
wanted title for these view pages and inject the title in the correct
region.