RTS-15298 Updated tabs javascript to new version of jquery
The tabs javascript was based on the default jquery version of drupal.
After installing jquery_update we use a newer version of jquery core and
ui. Some code used was deprecated. This moves the tabs script up to
date.

In persons and area the tabs dis not work as expected. This adds the
javascript to all use cases.