RTS-15497 Remove filetopic_ part from urls in text fields
At a time in w2 it was possible to find urls for a filetopic and make a
link to the topic, not the file it would represent. As long as w2 was
alive the filetopic would return the file, but when w2 was turned off
these urls stopped to return results. It should be safe to remove any
occurrence of filetopic_ from these texts.