RTS-15612 Implement logic to support a different image layout in article
We want the main media to get a different layout when a description is
attached to the image file. This is implemented by using
hook_preprocess_field to insert a class indicating the presence of a
description and by adding a container to the media information.

Includes styling done in commit:d70f6f7 in SG