RTS-15670 Redirect after messages are created or saved
We don't want to display the regular node view page for message
objects, so we redirect to the view for the corresponding area.
When a new message is created it should display at the top.

A complication is that we need to look up the area node matching
the current language, while merging this with the code to look
up the tnid for the area.