RTS-13482 Queries to remove area menu items with depth larger than 2
When adding a menu lint from the admin menu it has been (and is) possible to
select a parent item at level two. Though theses menu items are not
displayed it causes the parent items to pick up the class collapsed and
the default drupal styling attached to the class. Theses two queries
removes menu items in area menus with a depth larger than 2 and resets
the has_children value of all menu items at level 2.