RTS-16011 Update the /nb/<area>/meldinger.xml alias
This change makes the /nb/<area>/meldinger alias point to the Norwegian
version of the RSS view.  The old configuration made both alias forward
the request to the English view.

In the post-upgrade script all we have to do is enable it for the
current release.  The call to the area-aliases script was already
present.