RTS-15958 Make date icon show for small media in teaser
When no main media is present we use small media (if any) in article
teasers. This causes the render array to be a bit different and when
inserting the markup for the date icon it does not apply for the small
image. This should make the render arrays have the same structure.