RTS-16015 Hide not wanted position title
The position title for the position with position code 9300 is "Stilling
utenfor reg". This is not a position title we want to display in our
user profiles. We should hide this title and only show the alt position title
if this exists. This should apply to user pages, user teasers and user
vcards.