RTS-16238 Using a proper registered UiB app in dataporten
We should not use the demo app provided by yours truly. When users grant
permission to dataporten to certain information it looks better to
display the id of our organization. This will use the proper application
for www.uib.no and www.test.uib.no. We will still use the demo app for
development purposes for localhost:3000.