RTS-16320 Added option to show older messages
Currently we only show the last 12 messages or messages for the last
week. We want an option to show older messages. This is implemented by
adding an offset which considers messages already shown and a call to
fetch the remaining messages from the message api. The work also
includes some refactoring of the javascript code.

This commit includes the code added in SG commit:4a231db