RTS-16344 Added validation for field collection in content list step
At present it is possible to save a content list with an incorrect step
content. The process throws an error, but the node is saved and renders
the content list inaccessible to all users except user 1. This
validation is meant to insure that the step content has a numeric value
(node id).