RTS-16617 Sync over research agreements
These are are found in fs-pres as exchange agreements tagged with
category 'MOU' or 'FORSKNINGSAVTALE'.

We store the category as uib_study_category on the w3 side, which we can
use to know the difference between regular student exchange agreements
and these research agreements.

We lowercase the category string in w3 to make them take on the same
form as the the other categories we have for regular studies and delare
them in the content type definition so that we can filter them out
from the existing views.