RTS-16546 Prevent deleting nodes that are part of a content list
If a user deletes a node that is used as a step in a content list the
content list is no longer available for editing or deleting for anyone
but user 1. This fix hook into the delete confirm form and removes the
possibility to verify the deletion if the node is part of a content
list. The user is displayed a message and a suggested action if the node
really should be deleted.