RTS-16209 Added markdown to external author field
A request has been made to make it possible to add a url to authors
added to an article in the external author field. This is implemented by
adding a new textformat which makes it possible to use markdown in this
field. The only option supported is the markdown syntax to create a
link.